All backend features ready to go! - V1.02.z3

+++ NissanPractice/Models/Products.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public double ProductMass { get; set; }
-        public double ProductCost { get; set; }
+        public decimal ProductMass { get; set; }
+        public decimal ProductCost { get; set; }
     }
 }
